requirements.txt
Django==3.1.1
djangorestframework==3.11.1
pandas==1.1.2
scikit-learn==0.23.2
numpy==1.19.1
mysql-connector-python==8.0.21
mysqlclient==2.0.1
PyMySQL==0.10.1
SQLAlchemy==1.3.19